--- app/src/main/java/com/example/egobook_frontent/ui/diary/DiaryCheckFragment.kt
@@ -1,27 +1,0 @@
-package com.example.egobook_frontent.ui.diary
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.egobook_frontent.databinding.FragmentDiaryCheckBinding
-
-class DiaryCheckFragment : Fragment() {
-
-    private var _binding: FragmentDiaryCheckBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentDiaryCheckBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
