+++ app/src/main/java/com/example/egobook_frontent/ui/diary/view/DiaryCheckFragment.kt
@@ -1,0 +1,43 @@
+package com.example.egobook_frontent.ui.diary.view
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
+import com.example.egobook_frontent.databinding.FragmentDiaryCheckBinding
+
+class DiaryCheckFragment : Fragment() {
+
+    private var _binding: FragmentDiaryCheckBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentDiaryCheckBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        // 하단 시스템 바 영역만큼 패딩을 주어 버튼이 가려지지 않게 함.
+        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            // 기존 패딩은 유지하면서 하단만 시스템 바 높이만큼 추가
+            v.setPadding(v.paddingLeft, v.paddingTop, v.paddingRight, systemBars.bottom)
+            insets
+        }
+
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
