+++ app/src/main/res/color/selector_able_disable.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/cos_black" android:state_enabled="true"/>
+    <item android:color="@color/neutral_weak" android:state_selected="false"/>
+</selector>
