+++ app/src/main/res/layout/fragment_diary_write.xml
@@ -348,6 +348,6 @@
         android:fontFamily="@font/arita_semibold"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
+        app:layout_constraintEnd_toEndOf="parent"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
